etl/requirements.txt
boto3==1.40.49
botocore==1.40.49
certifi==2025.10.5
charset-normalizer==3.4.3
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
